Fix race-condition leading to missing DATA.
When moving push-promises handlers to streams (in #13 - 02b0552) I've
introduced a race condition because the client stream and the HPACK
stream do not have the same index on the server-sent frames channels.

This patch ensures that the push-promise stream starts with the frames
starting right after the PUSH_PROMISE is received.