Throw an exception on a GOAWAY.
This exception is thrown on the control-frames loop (an Async). If
callers linked against this async then the exception will be rethrown.
There is a race-condition risk if a GOAWAY is received before clients
have linked against the control-frames loop.

Also take advantage of the diff to correct a call to throw with throwIO.