Cleans up stack resolvers to latest per GHC
This change just moves all of the resolvers to the latest version
in their particular GHC series. Some of them still remain missing
and can be added in subsequent commits if people care. At the time
of this writing though the missing ones are at least a year old.

Also adds a test script to test all of the builds that are supported
in the repository.

The main stack.yaml continues to point to the oldest supported
version so that compatibility isn't broken invisibly.